moved content from parser pkg to indexer, removed parser pkg.
