+++ src/indexer/Parser.ts
@@ -1,4 +1,4 @@
-import { IsInIndexItem} from '../IsInIndexItem';
+import { IsInIndexItem} from './IsInIndexItem';
 import { ParserErrorEmptyJSONString, ParserErrorMissingProperties } from './Error';
 
 interface JSONBody {
@@ -8,20 +8,19 @@ interface JSONBody {
 }
 
 class Parser {
-  static hasUndefinedProperties(obj: JSONBody): boolean {
+  private static hasUndefinedProperties(obj: JSONBody): boolean {
     const missingISIN = obj.isin === undefined;
     const missingFigi = obj.figi ===  undefined;
     const missingCompanyName = obj.companyname === undefined;
     return missingISIN || missingFigi ||  missingCompanyName;
   }
 
-  static ToIsInIndexItem(_jsonText: string): IsInIndexItem {
-    // validation will be extracted to separate private func
-    if (!_jsonText.length)
+  public ConvertToIsInIndexItem(jsonText: string): IsInIndexItem {
+    if (!jsonText.length)
       throw ParserErrorEmptyJSONString;
 
-    const obj: JSONBody = JSON.parse(_jsonText);
-    const missingProperties: boolean = this.hasUndefinedProperties(obj)
+    const obj: JSONBody = JSON.parse(jsonText);
+    const missingProperties: boolean = Parser.hasUndefinedProperties(obj)
     if (missingProperties) {
       throw ParserErrorMissingProperties;
     }
