+++ src/indexer/SearchIndexer.test.ts
