+++ src/indexer/SearchIndexer.ts
